requirements.txt
click==8.1.3
pydantic==1.10.2
pyyaml==6.0
tqdm==4.64.1
torch==1.13.0+cu116
torchvision==0.14.0+cu116
torchaudio==0.13.0+cu116
tensorboard==2.8.0
GPUtil==1.4.0
cuda-python==11.8.1
numpy==1.23.5
matplotlib==3.6.2
soundfile==0.11.0
